API reference

SecureArithmetic.PlainVector — Type

PlainVector

Holds encoded - but not encrypted - data for arithmetic operations. Can be converted to a SecureVector using encrypt.

See also: SecureVector, encrypt

SecureArithmetic.SecureVector — Type

SecureVector

Holds encrypted data for arithmetic operations. Can be converted to a PlainVector using decrypt.

See also: PlainVector, decrypt

Base.circshift — Method

circshift(sv::SecureVector, shift; wrap_by = :capacity)

Circularly shift, i.e., rotate the data in sv by shift positions, similarly to Julia's circshift for regular arrays. wrap_by indicates whether the rotation should be applied with respect to the current data length of sv (wrap_by :length) or with respect to its maximum capacity (wrap_by = :capacity).

Note: If sv's length is less than its capacity, wrapping by :length increases the multiplicative depth of your algorithm by one and is more expensive to compute. Furthermore, one additional rotation is applied with a shift of -sign(shift) * (length(sv) - abs(shift)).

See also: SecureVector, length, capacity

Base.length — Method

length(v::Union{PlainVector, SecureVector})

Return the current length of v, i.e., the number of container elements in use. Note that this might be less than its maximum capacity.

See also: capacity, SecureVector, PlainVector

SecureArithmetic.capacity — Method

capacity(v::Union{PlainVector, SecureVector})

Return the current capacity of v, i.e., the maximum number of elements the container may hold.. Note that this might be more than its current length.

See also: length, SecureVector, PlainVector

SecureArithmetic.decrypt! — Method

decrypt!(plain_vector::PlainVector, secure_vector::SecureVector, private_key::PrivateKey)

Decrypt secure_vector using the private_key and store the result in the given plain_text.

See also: PlainVector, SecureVector, decrypt

SecureArithmetic.decrypt — Method

decrypt(secure_vector::SecureVector, private_key::PrivateKey)

Decrypt secure_vector using the private_key and return the resulting PlainVector.

See also: PlainVector, SecureVector, decrypt!

SecureArithmetic.encrypt — Method

encrypt(plain_vector::PlainVector, public_key::PublicKey)

Encrypt plain_vector into a SecureVector using the public_key.

See also: PlainVector, SecureVector, decrypt

SecureArithmetic.encrypt — Method

encrypt(data::Vector{<:Real}, public_key::PublicKey, context::SecureContext)

Encrypt data into a SecureVector using the public_key derived for the given context.

See also: SecureVector, decrypt