API reference

SecureArithmetic.OpenFHEBackend — Type

OpenFHEBackend

Cryptography backend for use with the homomorphic encryption library OpenFHE (https://github.com/openfheorg/openfhe-development).

See also: SecureContext, Unencrypted

SecureArithmetic.PlainVector — Type

PlainVector

Holds encoded - but not encrypted - data for arithmetic operations. Can be converted to a SecureVector using encrypt.

See also: SecureVector, encrypt

SecureArithmetic.PlainVector — Method

PlainVector(data::Vector{<:Real}, context::SecureContext{<:OpenFHEBackend})

Constructor for data type PlainVector takes an unencrypted data vector and a context object of type SecureContext{<:OpenFHEBackend}. Return PlainVector with encoded but not encrypted data. The context can be utilized later for encryption using encrypt, resulting in SecureVector.

See also: PlainVector, SecureVector, encrypt, decrypt OpenFHEBackend

SecureArithmetic.PlainVector — Method

PlainVector(data::Vector{<:Real}, context::SecureContext{<:Unencrypted})

Constructor for data type PlainVector takes an unencrypted data vector and a context object of type SecureContext{<:Unencrypted}. Returns PlainVector with not encoded and not encrypted data. The context can be utilized later for encryption using encrypt, resulting in SecureVector.

See also: PlainVector, SecureVector, encrypt, decrypt

SecureArithmetic.PlainVector — Method

PlainVector(data::Vector{Float64}, context::SecureContext{<:OpenFHEBackend})

Constructor for data type PlainVector takes an unencrypted data vector and a context object of type SecureContext{<:OpenFHEBackend}. Return PlainVector with encoded but not encrypted data. The context can be utilized later for encryption using encrypt, resulting in SecureVector.

See also: PlainVector, SecureVector, encrypt, decrypt OpenFHEBackend

SecureArithmetic.PrivateKey — Type

PrivateKey

Holds a private key that is used for decryption in decrypt.

See also: PlainVector, SecureVector, decrypt

SecureArithmetic.PublicKey — Type

PublicKey

Holds a public key that is used for encryption in encrypt.

See also: PlainVector, SecureVector, encrypt

SecureArithmetic.SecureContext — Type

SecureContext

A structure used to generalize CryptoContext defined in OpenFHE.jl for unencrypted data, to maximize utilization of the same code for both plaintext and ciphertext.

See also: OpenFHEBackend, Unencrypted

SecureArithmetic.SecureVector — Type

SecureVector

Holds encrypted data for arithmetic operations. Can be converted to a PlainVector using decrypt.

See also: PlainVector, decrypt

SecureArithmetic.Unencrypted — Type

Unencrypted

An alternative backend to use instead of OpenFHEBackend to experiment with algorithms on unencrypted data.

See also: SecureContext, OpenFHEBackend

Base.circshift — Method

circshift(sv::SecureVector, shift; wrap_by = :capacity)

Circularly shift, i.e., rotate the data in sv by shift positions, similarly to Julia's circshift for regular arrays. wrap_by indicates whether the rotation should be applied with respect to the current data length of sv (wrap_by :length) or with respect to its maximum capacity (wrap_by = :capacity).

Note: If sv's length is less than its capacity, wrapping by :length increases the multiplicative depth of your algorithm by one and is more expensive to compute. Furthermore, one additional rotation is applied with a shift of -sign(shift) * (length(sv) - abs(shift)).

See also: SecureVector, length, capacity

Base.collect — Method

collect(v::PlainVector{<:OpenFHEBackend})

Decode and return the real-valued data contained in v.

See also: PlainVector, OpenFHEBackend

Base.collect — Method

collect(v::PlainVector{<:Unencrypted})

Return the real-valued data contained in v.

See also: PlainVector

Base.length — Method

length(v::Union{PlainVector, SecureVector})

Return the current length of v, i.e., the number of container elements in use. Note that this might be less than its maximum capacity.

See also: capacity, SecureVector, PlainVector

SecureArithmetic.bootstrap! — Method

bootstrap!(secure_vector::SecureVector{<:OpenFHEBackend})

Refresh a given secure_vector to increase the multiplication depth. Supported for CKKS only.

See also: SecureVector, OpenFHEBackend, init_bootstrapping!

SecureArithmetic.bootstrap! — Method

bootstrap!(secure_vector::SecureVector{<:Unencrypted})

An empty duplicate of bootstrap! for unencrypted data.

See also: SecureVector, Unencrypted, bootstrap!, init_bootstrapping!

SecureArithmetic.capacity — Method

capacity(v::Union{PlainVector, SecureVector})

Return the current capacity of v, i.e., the maximum number of elements the container may hold.. Note that this might be more than its current length.

See also: length, SecureVector, PlainVector

SecureArithmetic.decrypt! — Method

decrypt!(plain_vector::PlainVector, secure_vector::SecureVector, private_key::PrivateKey)

Decrypt secure_vector using the private_key and store the result in the given plain_text.

See also: PlainVector, SecureVector, decrypt

SecureArithmetic.decrypt — Method

decrypt(secure_vector::SecureVector, private_key::PrivateKey)

Decrypt secure_vector using the private_key and return the resulting PlainVector.

See also: PlainVector, SecureVector, decrypt!

SecureArithmetic.encrypt — Method

encrypt(plain_vector::PlainVector, public_key::PublicKey)

Encrypt plain_vector into a SecureVector using the public_key.

See also: PlainVector, SecureVector, decrypt

SecureArithmetic.encrypt — Method

encrypt(data::Vector{<:Real}, public_key::PublicKey, context::SecureContext)

Encrypt data into a SecureVector using the public_key derived for the given context.

See also: SecureVector, decrypt

SecureArithmetic.generate_keys — Method

generate_keys(context::SecureContext{<:OpenFHEBackend})

Generate and return public and private keys.

See also: PublicKey, PrivateKey, SecureContext, OpenFHEBackend

SecureArithmetic.generate_keys — Method

generate_keys(context::SecureContext{<:Unencrypted})

Return public and private keys for use with unencrypted data.

See also: PublicKey, PrivateKey, SecureContext, Unencrypted

SecureArithmetic.get_crypto_context — Method

get_crypto_context(context::SecureContext{<:OpenFHEBackend})

Return a OpenFHE.CryptoContext object stored in a given context.

See also: SecureContext, OpenFHEBackend

SecureArithmetic.get_crypto_context — Method

get_crypto_context(v::Union{SecureVector{<:OpenFHEBackend},
                            PlainVector{<:OpenFHEBackend},
                            SecureMatrix{<:OpenFHEBackend},
                            PlainMatrix{<:OpenFHEBackend}})

Return a OpenFHE.CryptoContext object stored in v.

See also: SecureContext, SecureVector, PlainVector, OpenFHEBackend

SecureArithmetic.init_bootstrapping! — Method

init_bootstrapping!(context::SecureContext{<:OpenFHEBackend},
                    private_key::PrivateKey)

Generate the necessary keys from private_key to enable bootstrapping for a given context. Supported for CKKS only.

See also: SecureContext, OpenFHEBackend, PrivateKey, bootstrap!

SecureArithmetic.init_bootstrapping! — Method

init_bootstrapping!(context::SecureContext{<:Unencrypted}, private_key::PrivateKey)

An empty duplicate of init_bootstrapping! for unencrypted data.

See also: SecureContext, Unencrypted, PrivateKey, init_bootstrapping!

SecureArithmetic.init_matrix_rotation! — Method

init_matrix_rotation!(context::SecureContext{<:OpenFHEBackend}, private_key::PrivateKey,
                      shifts, shape)

Generate rotation keys for matrix rotation with OpenFHE.EvalRotate using the private_key and for the rotation indexes in shifts. shifts is a pair of integers given as a tuple, or a list of tuples of integers. The keys are stored in the given context. A positive shift defines a rotation to the right/bottom, e.g., a rotation with shift (1, 0): [1 2 3; 4 5 6; 7 8 9] -> [7 8 9; 1 2 3; 4 5 6]. Negative shifts define rotation to the left/top, e.g., a rotation with a shift (0, -1): [1 2 3; 4 5 6; 7 8 9] -> [3 1 2; 6 4 5; 9 7 8].

SecureArithmetic.init_multiplication! — Method

init_multiplication!(context::SecureContext{<:OpenFHEBackend},
                     private_key::PrivateKey)

Generate relinearization key for use with OpenFHE.EvalMult using the private_key, and store it in the given context.

See also: SecureContext, OpenFHEBackend, PrivateKey

SecureArithmetic.init_multiplication! — Method

init_multiplication!(context::SecureContext{<:Unencrypted}, private_key::PrivateKey)

An empty duplicate of init_multiplication! for unencrypted data.

See also: SecureContext, Unencrypted, PrivateKey, init_multiplication!

SecureArithmetic.init_rotation! — Method

init_rotation!(context::SecureContext{<:OpenFHEBackend}, private_key::PrivateKey,
               shifts)

Generate rotation keys for use with OpenFHE.EvalRotate using the private_key and for the rotation indexes in shifts. The keys are stored in the given context. Positive shift defines rotation to the right, e.g. a rotation with a shift 1: [1, 2, 3, 4] -> [4, 1, 2, 3]. Negative shift defines rotation to the left, e.g. a rotation with a shift -1: [1, 2, 3, 4] -> [2, 3, 4, 1].

Note: Here, indexes stored in shifts have reversed sign compared to rotation indexes used in OpenFHE.

See also: SecureContext, OpenFHEBackend, PrivateKey

SecureArithmetic.init_rotation! — Method

init_rotation!(context::SecureContext{<:Unencrypted}, private_key::PrivateKey, shifts)

An empty duplicate of init_rotation! for unencrypted data.

See also: SecureContext, Unencrypted, PrivateKey, init_rotation!

SecureArithmetic.level — Method

level(v::Union{SecureVector{<:OpenFHEBackend}, PlainVector{<:OpenFHEBackend}})

Return the number of scalings, referred to as the level, performed over v.

See also: PlainVector, SecureVector, OpenFHEBackend

SecureArithmetic.level — Method

level(v::Union{SecureVector{<:Unencrypted}, PlainVector{<:Unencrypted}})

Return the number of scalings, referred to as the level, performed over v. For data type derived from Unencrypted, the level is always equal to 0.

See also: PlainVector, SecureVector

SecureArithmetic.release_context_memory — Method

release_context_memory()

Release all OpenFHE.CryptoContexts and keys for multiplication, rotation, bootstrapping and OpenFHE.EvalSum generated in the functions init_multiplication!, init_rotation!, init_bootstrapping! and OpenFHE.EvalSumKeyGen.

In the source code of OpenFHE C++, all CryptoContexts and keys are stored in static objects. Without using release_context_memory, the memory allocated for these contexts and keys will only be freed after restarting the Julia REPL. It is also advisable to call GC.gc() after a call to release_context_memory to clean up all memory on the Julia side.

See also: init_multiplication!, init_rotation!, init_bootstrapping!